# backend/docs/openapi/api.yaml
...
openapi: 3.0.3
info:
    description: |
        Machine-generated OpenAPI specification for AppOS APIs, with ext routes generated from route code using group-matrix as single source of truth.
    title: AppOS API
    version: 0.1.0
...
components:
    schemas:
        ErrorEnvelope:
            properties:
                code:
                    type: integer
                message:
                    type: string
            required:
                - message
            type: object
        GenericRequest:
            additionalProperties: true
            description: Generic request payload placeholder (refine per endpoint)
            type: object
...
        SuccessEnvelope:
            additionalProperties: true
            description: Generic success payload envelope
            example:
                ok: true
            type: object
    securitySchemes:
        bearerAuth:
            bearerFormat: PocketBase token
            scheme: bearer
            type: http
paths:
...
    /api/backups/{name}/restore:
        post:
            operationId: pb_backups_restore
            parameters:
                - in: path
                  name: name
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
            security:
                - bearerAuth: []
            summary: Restore backup by name
            tags:
                - Backups
...
    /api/ext/backup/create:
        post:
            description: Enqueues an asynchronous task to create a named backup of the platform state. Superuser only.
            operationId: post_api_ext_backup_create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenericRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "202":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: Accepted
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "503":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Service Unavailable
            security:
                - bearerAuth: []
            summary: Enqueue backup creation
            tags:
                - Backups
    /api/ext/backup/list:
        get:
            description: Returns all available backup files from platform storage. Superuser only.
            operationId: get_api_ext_backup_list
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: List available backups
            tags:
                - Backups
    /api/ext/backup/restore:
        post:
            description: Restores platform state from a named backup synchronously. Superuser only.
            operationId: post_api_ext_backup_restore
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenericRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Restore from backup
            tags:
                - Backups
...
    /api/ext/proxy/domains:
        get:
            description: Returns all active reverse proxy domain bindings. Superuser only.
            operationId: get_api_ext_proxy_domains
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: List proxy domains
            tags:
                - Proxy
        post:
            description: Creates a new reverse proxy domain binding (virtual host). Superuser only.
            operationId: post_api_ext_proxy_domains
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenericRequest'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Add proxy domain
            tags:
                - Proxy
    /api/ext/proxy/domains/{domain}:
        delete:
            description: Deletes a reverse proxy domain binding. Superuser only.
            operationId: delete_api_ext_proxy_domains_domain
            parameters:
                - in: path
                  name: domain
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Remove proxy domain
            tags:
                - Proxy
    /api/ext/proxy/domains/{domain}/ssl:
        post:
            description: Triggers SSL certificate provisioning for the given domain. Superuser only.
            operationId: post_api_ext_proxy_domains_domain_ssl
            parameters:
                - in: path
                  name: domain
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenericRequest'
                required: false
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Bad Request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Request SSL certificate
            tags:
                - Proxy
    /api/ext/proxy/reload:
        post:
            description: Sends a reload signal to the reverse proxy, applying pending configuration changes. Superuser only.
            operationId: post_api_ext_proxy_reload
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GenericRequest'
                required: false
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessEnvelope'
                    description: OK
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Unauthorized
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorEnvelope'
                    description: Internal Server Error
            security:
                - bearerAuth: []
            summary: Reload proxy config
            tags:
                - Proxy
...